Update README to point to authoritative docs.
The README was referencing an outdated source of information and
included many details that had become stale over time, as this wasn't
being updated when new prebuilts were provided.

Refer instead to the current authoritative documentation describing how
to build WebView for AOSP, and remove the specific details about how to
rebuild the prebuilt as they're prone to becoming outdated and are not
really necessary in any case - rebuilding the same version is not
typically desirable and the Chromium docs suggest appropriate build
arguments for the current version.

Fixes: 139300536
Test: n/a, doc change only
Change-Id: I486c6c436d77d021b36b3618fe763f1bbe7079c0
(cherry picked from commit 6174fba679427a2c141036089253efd48c6b0371)