Remove undesirable JNI lib list in makefile.
Cherrypick of ag/10846513
We were listing the JNI libraries explicitly with
LOCAL_PREBUILT_JNI_LIBS in the WebView prebuilt makefile - this hasn't
been necessary for some time, and actually causes any other .so in the
APK to be removed at build time, which is now removing the crashpad
trampoline .so and probably causing crashpad to not work correctly on
the preinstalled version as a result.

Just drop it entirely - the .so files are all preserved in the APK by
default.

Fixes: 152435891
Test: m; examine /product/app/webview/webview.apk
Change-Id: I047c0aedc4b332897a32a7f3a75a036d36b32bc1